# pubky-homeserver/openapi.yml
openapi: 3.1.0
info:
  title: Pubky Homeserver API
  description: |
    The Pubky Homeserver exposes three servers:
    - **Client Server** — Primary API for user authentication, file storage, and event streaming.
    - **Admin Server** — Password-protected administration endpoints.
    - **Metrics Server** — Prometheus metrics (`GET /metrics`).

    This specification covers the Client Server and Admin Server.

    ## Tenant Identification

    Per-tenant (user) routes resolve the target user from these sources (highest priority first):
    1. `pubky-host` header (explicit override; takes precedence over `Host`)
    2. `Host` header (e.g. from TLS SNI on the Pubky socket)
    3. `?pubky-host=` query parameter (fallback when no key found in headers)

    The value must be a z-base-32 encoded Ed25519 public key.

    ## Authentication

    Two authentication methods exist:
    - **Grant-based (current):** `Authorization: Bearer <token>` obtained via `POST /auth/grant/session`.
    - **Cookie-based (deprecated):** Session cookie set by `POST /signup` or `POST /session`.

    Grant middleware runs first; if it produces a session, cookie middleware is skipped.

    ## Capabilities

    Write operations are gated by capability-based access control. Capabilities are
    path-scoped (e.g. `read_write:/pub/app/`) or `root` (full access to `/pub/`).

    ## Write Path Restriction

    All PUT/DELETE operations require paths under `/pub/`. Attempts to write elsewhere
...
  version: 0.9.0
...
servers:
...
paths:
...
  /events/:
    get:
      tags: [Events]
      summary: Historical event feed (plain text)
      description: |
        Legacy endpoint returning a plain-text feed of file change events.
      operationId: getEventFeed
      parameters:
        - name: cursor
          in: query
          description: Starting cursor position. Defaults to `"0"` (beginning) when omitted.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of events to return.
          schema:
            type: integer
            minimum: 1
            maximum: 65535
      responses:
        "200":
          description: Event feed
          content:
            text/plain:
              schema:
                type: string
              example: |
                PUT pubky://o1gg96ewuo.../pub/example.txt
                DEL pubky://o1gg96ewuo.../pub/old.txt
                cursor: 12345
        "400":
          description: Invalid cursor
  /events-stream:
    get:
      tags: [Events]
      summary: Real-time event stream (SSE)
      description: |
        Server-Sent Events endpoint with two modes:
        - **Batch mode** (`live=false` or omitted): fetches historical events then closes.
        - **Live mode** (`live=true`): fetches historical events, then streams new events in real-time.

        Slow clients that fall behind in live mode will have their connection closed.

        Each SSE message has an event type (`PUT` or `DEL`) and multiline data:
        ```
        event: PUT
        data: pubky://user_pubkey/pub/example.txt
        data: cursor: 42
        data: content_hash: r0NJufX5oaagQE3qNtzJSZvLJcmtwRK3zJqTyuQfMmI=
        ```
      operationId: getEventStream
      parameters:
        - name: user
          in: query
          required: true
          description: |
            One or more user public keys (z-base-32). Repeat for multiple users.
            Optionally append `:cursor` to start from a specific position.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          examples:
            single:
              value: [o1gg96ewuojmopcjbz8895478wdtxtzzuxnfjjz8o8e77csa1ngo]
            withCursor:
              value: ["o1gg96ewuojmopcjbz8895478wdtxtzzuxnfjjz8o8e77csa1ngo:42"]
        - name: limit
          in: query
          description: Maximum total events to send.
          schema:
            type: integer
            minimum: 1
            maximum: 65535
        - name: reverse
          in: query
          description: Reverse chronological order. Cannot be combined with `live=true`.
          schema:
            type: boolean
            default: false
        - name: live
          in: query
          description: Enable live streaming mode. Cannot be combined with `reverse=true`.
          schema:
            type: boolean
            default: false
        - name: path
          in: query
          description: Path prefix filter (e.g. `/pub/files/` or `pub/files/`).
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: |
            Invalid user pubkey, missing user parameter, too many users,
            live + reverse combination, or invalid path.
        "404":
          description: User not found

  # ──────────────────────────────────────────────────────────────────────
  # Admin Server
  # ──────────────────────────────────────────────────────────────────────
...
  /events:
    get:
      tags: [Admin]
      summary: All events feed (plain text, includes private paths)
      description: |
        Same plain-text, cursor-paginated format as the public `GET /events/` feed, but exposes
        all events including private (`/priv/...`) paths. Response is `Cache-Control: no-store`.
      operationId: getAdminEventFeed
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - name: cursor
          in: query
          description: Starting cursor position. Defaults to `"0"` (beginning) when omitted.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of events to return.
          schema:
            type: integer
            minimum: 1
            maximum: 65535
      responses:
        "200":
          description: Event feed (public and private events)
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            text/plain:
              schema:
                type: string
              example: |
                PUT pubky://o1gg96ewuo.../pub/example.txt
                PUT pubky://o1gg96ewuo.../priv/app/secret.txt
                DEL pubky://o1gg96ewuo.../pub/old.txt
                cursor: 12345
        "400":
          description: Invalid cursor
        "401":
          description: Missing or invalid admin password
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Grant-based bearer token from `POST /auth/grant/session`.

    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Deprecated. In practice, the cookie name is the user's z-base-32
        public key and the value is the session secret. The `name` field here
        is a placeholder; the actual name varies per user.

    adminPassword:
      type: apiKey
      in: header
      name: X-Admin-Password
      description: Admin server password.

    adminBasicAuth:
      type: http
      scheme: basic
      description: Basic auth for WebDAV endpoints. Username `admin`, password is the admin password.
...